Convert client test to jest
